client/backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
pymongo==4.6.0
motor==3.3.2
python-multipart==0.0.6
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-dotenv==1.0.0
pydantic==2.5.0
pydantic-settings==2.1.0
requests==2.31.0
google-generativeai==0.3.2
chromadb==0.4.18
sentence-transformers==2.2.2
numpy==1.24.3
websockets==15.0.1
python-socketio==5.13.0
PyJWT==2.8.0 